requirements.txt
requests
pysha3
